Add files via upload
This repository contains the results of our COVID-19 data analysis conducted during our intra-department hackathon. We utilized a combination of machine learning techniques, employing the Random Forest algorithm for model creation and Linear Regression for trend analysis.

Our analysis aims to provide insights into the spread and trends of COVID-19 using real-world data. By leveraging Random Forest, we aimed to develop a predictive model that could forecast potential infection rates and identify key factors influencing the spread of the virus. Additionally, our implementation of Linear Regression facilitated the exploration of temporal trends, aiding in understanding the progression of the pandemic over time.

This repository includes the code, datasets, and documentation detailing our methodology and findings. Through this project, we aim to contribute to the ongoing efforts in understanding and mitigating the impact of COVID-19.